% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigHotSpotter_pipeline.R
\name{SigHotSpotter_pipeline}
\alias{SigHotSpotter_pipeline}
\title{General pipeline for SigHotSpotter}
\usage{
SigHotSpotter_pipeline(species, input_data, cutoff, DE_Genes_data,
  percentile, invert_DE = FALSE, showprogress = TRUE)
}
\arguments{
\item{species}{Currently supported species: "HUMAN", "MOUSE"}

\item{input_data}{File name for input gene expression data}

\item{cutoff}{Maximum number of zero-value genes, above this cutoff the genes are excluded}

\item{DE_Genes_data}{Differential expression dataset (1 for up-regulated, -1 for down-regulated genes)}

\item{percentile}{Predicted intermediates are taken into account above this threshold}

\item{invert_DE}{If the differential expression should be inverted, default = FALSE}

\item{showprogress}{shows progress bar in shiny app if set to TRUE, set it to FALSE in batch mode without GUI, default = TRUE}
}
\value{
Compatibility scores
}
\description{
The function computes compatibility scores for signaling intermediates
}
