% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheSIG.R
\docType{package}
\name{NicheSIG-package}
\alias{NicheSIG}
\alias{NicheSIG-package}
\title{NicheSIG: A novel approach to infer the effect of the cellular environment (niche) on the cell signaling cascade}
\description{
This package introduces a Markov Chain Model for Cell Signaling Cascade Inference.
}
\details{
For more information see the paper "A computational approach to control cell phenotype for cellular rejuvenation strategies"
}
