% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheSIG.R
\docType{package}
\name{NicheSIG-package}
\alias{NicheSIG}
\alias{NicheSIG-package}
\title{NicheSIG: A computational approach to control cell phenotype for cellular rejuvenation strategies}
\description{
This package introduces a novel approach to infer the effect of the cellular environment (niche) on the cell signaling cascade
}
\details{
Markov Chain for Cell Signaling Cascade Inference
}
